+++ packages/@expo/cli/src/start/server/metro/__tests__/withMetroMultiPlatform.test.ts
@@ -100,19 +100,20 @@ describe(withExtendedResolver, () => {
     modified.resolver.resolveRequest!(getDefaultRequestContext(), 'react-native', platform);
 
     expect(getResolveFunc()).toBeCalledTimes(1);
-    expect(getResolveFunc()).toBeCalledWith(
+
+    expect(getResolveFunc()).toHaveBeenNthCalledWith(
+      1,
       expect.objectContaining({
-        nodeModulesPaths: ['/node_modules', '/src'],
         extraNodeModules: {},
         mainFields: ['react-native', 'browser', 'main'],
         preferNativePlatform: true,
       }),
-      'react-native',
+      '/src/react-native',
       platform
     );
   });
 
-  it(`resolves to react-native-web on web`, async () => {
+  it(`does not alias react-native-web in initial resolution with baseUrl on web`, async () => {
     mockMinFs();
 
     const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/' }), {
@@ -128,10 +129,33 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toBeCalledTimes(1);
     expect(getResolveFunc()).toBeCalledWith(
       expect.objectContaining({
-        nodeModulesPaths: ['/node_modules', '/src'],
-        extraNodeModules: {
-          'react-native': expect.stringContaining('node_modules/react-native-web'),
-        },
+        mainFields: ['browser', 'module', 'main'],
+        preferNativePlatform: false,
+      }),
+      '/src/react-native',
+      platform
+    );
+  });
+
+  it(`resolves to react-native-web on web`, async () => {
+    mockMinFs();
+
+    const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/' }), {
+      platforms: ['ios', 'web'],
+      tsconfig: {
+        // No baseUrl
+        paths: { '/*': ['*'] },
+      },
+      isTsconfigPathsEnabled: true,
+    });
+
+    const platform = 'web';
+
+    modified.resolver.resolveRequest!(getDefaultRequestContext(), 'react-native', platform);
+
+    expect(getResolveFunc()).toBeCalledTimes(1);
+    expect(getResolveFunc()).toBeCalledWith(
+      expect.objectContaining({
         mainFields: ['browser', 'module', 'main'],
         preferNativePlatform: false,
       }),
@@ -165,10 +189,6 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toBeCalledTimes(1);
     expect(getResolveFunc()).toBeCalledWith(
       expect.objectContaining({
-        nodeModulesPaths: ['/node_modules'],
-        extraNodeModules: {
-          'react-native': expect.stringContaining('node_modules/react-native-web'),
-        },
         mainFields: ['browser', 'module', 'main'],
         preferNativePlatform: false,
       }),
@@ -207,9 +227,6 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toBeCalledWith(
       expect.objectContaining({
         nodeModulesPaths: ['/node_modules'],
-        extraNodeModules: {
-          'react-native': expect.stringContaining('node_modules/react-native-web'),
-        },
         mainFields: ['browser', 'module', 'main'],
         preferNativePlatform: false,
       }),
@@ -243,9 +260,6 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toBeCalledTimes(1);
     expect(getResolveFunc()).toBeCalledWith(
       expect.objectContaining({
-        extraNodeModules: {
-          'react-native': expect.stringContaining('node_modules/react-native-web'),
-        },
         mainFields: ['main', 'module'],
         preferNativePlatform: false,
         // Moved mjs to the back